improve validation

+++ src/main/java/com/ctrip/zeus/service/model/handler/impl/AppSyncImpl.java
@@ -4,10 +4,11 @@ import com.ctrip.zeus.dal.core.*;
 import com.ctrip.zeus.exceptions.ValidationException;
 import com.ctrip.zeus.model.entity.*;
 import com.ctrip.zeus.service.model.handler.AppSync;
-import com.ctrip.zeus.service.model.handler.DbClean;
 import com.ctrip.zeus.support.C;
 import com.google.common.base.Function;
 import com.google.common.collect.Maps;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 import org.unidal.dal.jdbc.DalException;
 
@@ -35,42 +36,39 @@ public class AppSyncImpl implements AppSync {
     @Resource
     private SlbDao slbDao;
 
-    @Resource
-    private DbClean dbClean;
+    Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Override
     public AppDo add(App app) throws DalException, ValidationException {
-        if (!validate(app))
-            throw new ValidationException("app contains invalid information.");
+        validate(app);
         AppDo d= C.toAppDo(app);
         d.setCreatedTime(new Date());
         d.setVersion(1);
+
         appDao.insert(d);
-        sync(d, app);
+        cascadeSync(d, app);
 
         return d;
     }
 
     @Override
     public AppDo update(App app) throws DalException, ValidationException {
-        if (!validate(app)) {
-            throw new ValidationException("app contains invalid information.");
-        }
+        validate(app);
         AppDo d= C.toAppDo(app);
         appDao.updateByName(d, AppEntity.UPDATESET_FULL);
 
         AppDo updated = appDao.findByName(app.getName(), AppEntity.READSET_FULL);
         d.setId(updated.getId());
         d.setVersion(updated.getVersion());
-
-        sync(d, app);
+        cascadeSync(d, app);
         return d;
     }
 
     @Override
     public int delete(String name) throws DalException {
         AppDo d = appDao.findByName(name, AppEntity.READSET_FULL);
-
+        if (d == null)
+            return 0;
         appSlbDao.deleteByApp(new AppSlbDo().setAppName(d.getName()));
         appServerDao.deleteByApp(new AppServerDo().setAppId(d.getId()));
         appHealthCheckDao.deleteByApp(new AppHealthCheckDo().setAppId(d.getId()));
@@ -79,7 +77,17 @@ public class AppSyncImpl implements AppSync {
         return appDao.deleteByName(d);
     }
 
-    private boolean validate(App app) throws DalException {
+    private void validate(App app) throws DalException, ValidationException {
+        if (app == null) {
+            throw new ValidationException("App with null value cannot be persisted.");
+        }
+        if (!validateSlb(app))
+            throw new ValidationException("App with invalid slb data cannot be persisted.");
+    }
+
+    private boolean validateSlb(App app) throws DalException {
+        if (app.getAppSlbs().size() == 0)
+            return false;
         for (AppSlb as : app.getAppSlbs()) {
             if (slbDao.findByName(as.getSlbName(), SlbEntity.READSET_FULL) == null)
                 return false;
@@ -87,7 +95,7 @@ public class AppSyncImpl implements AppSync {
         return true;
     }
 
-    private void sync(AppDo d, App app) throws DalException {
+    private void cascadeSync(AppDo d, App app) throws DalException {
         syncAppSlbs(app.getName(), app.getAppSlbs());
         syncAppHealthCheck(d.getId(), app.getHealthCheck());
         syncLoadBalancingMethod(d.getId(), app.getLoadBalancingMethod());
@@ -116,23 +124,33 @@ public class AppSyncImpl implements AppSync {
 
         //Remove unused ones.
         for (AppSlbDo d : oldList) {
-            dbClean.deleteAppSlb(d.getId());
+            appSlbDao.deleteByPK(new AppSlbDo().setId(d.getId()));
         }
     }
 
     private void syncAppHealthCheck(long appKey, HealthCheck healthCheck) throws DalException {
+        if (healthCheck == null) {
+            logger.info("No health check method is found when adding/updating app with id " + appKey);
+            return;
+        }
         appHealthCheckDao.insert(C.toAppHealthCheckDo(healthCheck)
                 .setAppId(appKey)
                 .setCreatedTime(new Date()));
     }
 
     private void syncLoadBalancingMethod(long appKey, LoadBalancingMethod loadBalancingMethod) throws DalException {
+        if (loadBalancingMethod == null)
+            return;
         appLoadBalancingMethodDao.insert(C.toAppLoadBalancingMethodDo(loadBalancingMethod)
                 .setAppId(appKey)
                 .setCreatedTime(new Date()));
     }
 
     private void syncAppServers(long appKey, List<AppServer> appServers) throws DalException {
+        if (appServers == null || appServers.size() == 0) {
+            logger.warn("No app server is given when adding/update app with id " + appKey);
+            return;
+        }
         List<AppServerDo> oldList = appServerDao.findAllByApp(appKey, AppServerEntity.READSET_FULL);
         Map<String, AppServerDo> oldMap = Maps.uniqueIndex(oldList, new Function<AppServerDo, String>() {
             @Override
@@ -154,7 +172,7 @@ public class AppSyncImpl implements AppSync {
 
         //Remove unused ones.
         for (AppServerDo d : oldList) {
-            dbClean.deleteAppServer(d.getId());
+            appServerDao.deleteByPK(new AppServerDo().setId(d.getId()));
         }
     }
 }
